page_waterfall: fix visualization at 32-bits, refactor a bit
there was a lot of stuff that was simply unnecessary, like doing
*any* floating point calculation at all in getting the columns

also, I've lowered the FFT buffer size because at higher sizes
it would cause weird calculation errors at lower frequencies.
maybe it can be raised once I figure that crap out

+++ schism/util.c
@@ -41,61 +41,60 @@ extraneous libraries (i.e. GLib). */
 /* --------------------------------------------------------------------- */
 /* CONVERSION FUNCTIONS */
 
-/* linear -> decibel */
+/* linear -> deciBell */
 /* amplitude normalized to 1.0f. */
-double dB(double amplitude)
+float dB(float amplitude)
 {
-	return 20.0 * log10(amplitude);
+	return 20.0f * log10f(amplitude);
 }
 
-/* decibel -> linear */
-double dB2_amp(double db)
+/* deciBell -> linear */
+float dB2_amp(float db)
 {
-	return pow(10.0, db / 20.0);
+	return powf(10.0f, db / 20.0f);
 }
 
-/* linear -> decibel */
+/* linear -> deciBell */
 /* power normalized to 1.0f. */
-double pdB(double power)
+float pdB(float power)
 {
-	return 10.0 * log10(power);
+	return 10.0f * log10f(power);
 }
 
-/* decibel -> linear */
-double dB2_power(double db)
+/* deciBell -> linear */
+float dB2_power(float db)
 {
-	return pow(10.0, db / 10.0);
+	return powf(10.0f, db / 10.0f);
 }
-/* linear -> decibel
- * amplitude normalized to 1.0f.
- * Output scaled (and clipped) to 128 lines with noisefloor range.
- * ([0..128] = [-noisefloor..0dB])
- * correction_dBs corrects the dB after converted, but before scaling.
-*/
-short dB_s(int noisefloor, double amplitude, double correction_dBs)
+/* linear -> deciBell */
+/* amplitude normalized to 1.0f. */
+/* Output scaled (and clipped) to 128 lines with noisefloor range. */
+/* ([0..128] = [-noisefloor..0dB]) */
+/* correction_dBs corrects the dB after converted, but before scaling.*/
+short dB_s(int noisefloor, float amplitude, float correction_dBs)
 {
-	double db = dB(amplitude) + correction_dBs;
-	return CLAMP((int)(128.0*(db+noisefloor))/noisefloor, 0, 127);
+	float db = dB(amplitude) + correction_dBs;
+	return CLAMP((int)(128.f*(db+noisefloor))/noisefloor, 0, 127);
 }
 
-/* decibel -> linear */
+/* deciBell -> linear */
 /* Input scaled to 128 lines with noisefloor range. */
 /* ([0..128] = [-noisefloor..0dB]) */
 /* amplitude normalized to 1.0f. */
 /* correction_dBs corrects the dB after converted, but before scaling.*/
-short dB2_amp_s(int noisefloor, int db, double correction_dBs)
+short dB2_amp_s(int noisefloor, int db, float correction_dBs)
 {
-	return dB2_amp((db*noisefloor/128.0)-noisefloor-correction_dBs);
+	return dB2_amp((db*noisefloor/128.f)-noisefloor-correction_dBs);
 }
-/* linear -> decibel */
+/* linear -> deciBell */
 /* power normalized to 1.0f. */
 /* Output scaled (and clipped) to 128 lines with noisefloor range. */
 /* ([0..128] = [-noisefloor..0dB]) */
 /* correction_dBs corrects the dB after converted, but before scaling.*/
-short pdB_s(int noisefloor, double power, double correction_dBs)
+short pdB_s(int noisefloor, float power, float correction_dBs)
 {
 	float db = pdB(power)+correction_dBs;
-	return CLAMP((int)(128.0*(db+noisefloor))/noisefloor, 0, 127);
+	return CLAMP((int)(128.f*(db+noisefloor))/noisefloor, 0, 127);
 }
 
 /* deciBell -> linear */
@@ -103,7 +102,7 @@ short pdB_s(int noisefloor, double power, double correction_dBs)
 /* ([0..128] = [-noisefloor..0dB]) */
 /* power normalized to 1.0f. */
 /* correction_dBs corrects the dB after converted, but before scaling.*/
-short dB2_power_s(int noisefloor, int db, double correction_dBs)
+short dB2_power_s(int noisefloor, int db, float correction_dBs)
 {
 	return dB2_power((db*noisefloor/128.f)-noisefloor-correction_dBs);
 }
